added overlay option to control recursive directory overlay, specs for symlink and overlay

+++ lib/homesick/actions.rb
@@ -79,6 +79,7 @@ class Homesick
       source = Pathname.new(source)
       destination = Pathname.new(destination)
       recurse = false
+      
       if destination.symlink?
         if destination.readlink == source
           say_status :identical, destination.expand_path, :blue unless options[:quiet]
@@ -90,13 +91,13 @@ class Homesick
           end
         end
       elsif destination.exist?
-        if destination.directory? && source.directory?
+        if destination.directory? && source.directory? && options[:overlay]
           say_status :overlaying, "#{destination} exists, overlaying directory contents", :blue unless options[:quiet]
           recurse = true
         else
           say_status :conflict, "#{destination} exists", :red unless options[:quiet]
           if options[:force] || shell.file_collision(destination) { source }
-            if source.directory?
+            if source.directory? && options[:overlay]
               system "rm -rf #{destination}" unless options[:pretend]
               system "mkdir #{destination}" unless options[:pretend]
               recurse = true
@@ -107,7 +108,6 @@ class Homesick
         end
       else
         if source.directory?
-          #say_status :creating, "Created new directory at #{destination}", :blue unless options[:quiet]
           system "mkdir #{destination}" unless options[:pretend]
           recurse = true
         else
@@ -115,10 +115,12 @@ class Homesick
         end
       end
 
-      if recurse
-        files = Pathname.glob([source+"*", source+".*"]).reject{|a| [".",".."].include?(a.split.last.to_s)}          
+      puts BobsYourUncle      
+
+      if recurse && options[:overlay]
+        files = Pathname.glob([source+"*", source+".*"]).reject{|a| [".",".."].include?(a.split.last.to_s)}
         files.each do |path|
-          ln_s path.expand_path, destination.expand_path + path.split.last
+          self.ln_s path.expand_path, destination.expand_path + path.split.last
         end  
       end
     end
